% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy measures for a model}
\usage{
accuracy(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\description{
Returns range of summary measures of the forecast accuracy.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("accuracy")}
}

