% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_text.R
\name{alt_text}
\alias{alt_text}
\title{Alternative text generation}
\usage{
alt_text(x, ...)
}
\arguments{
\item{x}{A An object.}

\item{...}{Other arguments passed to methods}
}
\value{
A character string.
}
\description{
Return a character string that summarizes an object. This can be used to
improve the accessibility by passing the results to visualization and
documentation functions (e.g. the \code{fig.alt} chunk option in Rmarkdown
documents).
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("alt_text")}
}

