% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation}
\alias{annotation}
\alias{annotation<-}
\title{Accessing annotation information}
\usage{
annotation(object, ...)

annotation(object, ...) <- value
}
\arguments{
\item{object}{An object containing annotation information.}

\item{...}{Additional arguments, for use in specific methods.}

\item{value}{The annotation information to set on \code{object}.}
}
\description{
Get or set the annotation information contained in an object.
}
\examples{
annotation
showMethods("annotation")
}
\seealso{
\itemize{
   \item \code{\link[methods]{showMethods}} for displaying a summary of the
         methods defined for a given generic function.

   \item \code{\link[methods]{selectMethod}} for getting the definition of
         a specific method.

   \item \link{generics} for a summary of all the generics defined
         in the \pkg{generics} package.
 }
}
\keyword{methods}

