% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame}
\alias{as.data.frame}
\title{The as.data.frame() generic}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an R object}

\item{row.names}{NULL or a character vector giving the row names for
the data frame. Missing values are not allowed.}

\item{optional}{If TRUE, setting row names and converting column names
(to syntactic names: see \link[base]{make.names}) is optional. Note that all of R's
base package \code{as.data.frame()} methods use \code{optional} only for column names
treatment, basically with the meaning of}

\item{...}{additional arguments to be passed to or from methods.
\link[base]{data.frame}\code{(*, check.names = !optional)}.}
}
\description{
Generic for the \link[base]{as.data.frame} function
}
