% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\name{as.list}
\alias{as.list}
\title{The as.list() generic}
\usage{
as.list(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{objects, possibly names}
}
\description{
Generic for the \link[base]{list} function
}
