% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{augment}
\alias{augment}
\title{augment}
\usage{
augment(x, ...)
}
\arguments{
\item{x}{An object. See the individual method for specifics.}

\item{...}{Additional arguments to pass to other functions.}
}
\description{
augment
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("augment")}
}

