% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate statistics.}
\usage{
calculate(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods}
}
\description{
Calculate statistics.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("calculate")}
}

