% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.R
\name{cbind}
\alias{cbind}
\alias{rbind}
\title{The cbind() and rbind() generics}
\usage{
cbind(..., deparse.level = 1)

rbind(..., deparse.level = 1)
}
\arguments{
\item{...}{(generalized) vectors or matrices. These can be given as named arguments.
Other R objects may be coerced as appropriate, or S4 methods may be used: see
sections ‘Details’ and ‘Value’. (For the "data.frame" method of cbind these can
be further arguments to data.frame such as stringsAsFactors.)}

\item{deparse.level}{integer controlling the construction of labels in the case
of non-matrix-like arguments (for the default method):
deparse.level = 0 constructs no labels; the default,
deparse.level = 1 or 2 constructs labels from the argument names,
see the ‘Value’ section below.}
}
\description{
Generics for the \code{\link[base]{cbind}}/\code{rbind} functions
}
\note{
Need to explicitly define those generics otherwise the implicit generics
in package "base" would dispatch on 'deparse.level'.

Note that dispatching on '...' is supported starting with R 2.8.0 only.
}
