% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{coercion}
\alias{coercion}
\alias{as.factor}
\alias{as.ordered}
\alias{as.difftime}
\alias{as.difftime.default}
\title{Coercion}
\usage{
as.factor(x, ...)

as.ordered(x, ...)

as.difftime(tim, ...)

\method{as.difftime}{default}(tim, format = "\%X", units = "auto", ...)
}
\arguments{
\item{x}{A vector of data.}

\item{tim}{A vector specifying a time interval.}

\item{format}{A single character specifying the format of \code{tim} when it is
a character. The default is a locale-specific time format.}

\item{units}{A single character specifying units in which the results are
desired. Required if \code{tim} is a numeric.}
}
\value{
For \code{as.factor()} and \code{as.ordered()}, a factor or ordered factor.

For \code{as.difftime()}, a difftime object with an attribute indicating the
units.
}
\description{
Coercion functions (\code{as.*()}) for creating various classes from other
existing objects.
}
\details{
These functions override non-generic coercion functions provided in base
so that packages can provide methods for different data types. The
default methods call the base versions.
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("as.factor")}
}

\examples{
as.factor(letters[1:5])
as.ordered(letters[1:5])

as.difftime(1:5, units = "secs")
as.difftime(c("01:55:22", "01:55:25"))
as.difftime("01", format = "\%H")
as.difftime("01", format = "\%H", units = "secs")

}
