% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{compile}
\alias{compile}
\title{compile}
\usage{
compile(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
compile
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("compile")}
}

