% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components}
\alias{components}
\title{Extract components}
\usage{
components(object, ...)
}
\arguments{
\item{object}{A data separable object.}

\item{...}{Other arguments passed to methods}
}
\value{
A dataset (\code{\link[tibble:tibble]{tibble::tibble()}} or similar) containing components from
the object.
}
\description{
Decomposition methods and some modelling techniques can be used to decompose
a dataset into components of interest. This function is used to extract
these components in a tidy data format.
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("components")}
}

