% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R
\name{equation}
\alias{equation}
\title{Model equations}
\usage{
equation(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Other arguments passed to methods}
}
\value{
Markup output suitable for rendering the equation.
}
\description{
Display the mathematical representation of a fitted model.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("equation")}
}

