% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{explain}
\alias{explain}
\title{explain}
\usage{
explain(x, ...)
}
\arguments{
\item{x}{Model object or other R object with information to append to
observations.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
explain
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("explain")}
}

