% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Create an interactive visualization appropriate to a particular object type}
\usage{
explore(x, ...)
}
\arguments{
\item{x}{A object}

\item{...}{Other arguments passed to methods}
}
\value{
\code{NULL} (invisibly) or some other data type (e.g. tibble) depending on
the application.
}
\description{
\code{explore()} invokes a function that starts an interactive, pre-defined widget
(e.g. \code{plotly} visualization, \code{shiny} app, etc.) to investigate the results.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("explore")}
}

