% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{find_varying}
\alias{find_varying}
\title{find_varying}
\usage{
find_varying(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
find_varying
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("find_varying")}
}

