% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_xy}
\alias{fit_xy}
\title{fit_xy}
\usage{
fit_xy(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
fit_xy
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("fit_xy")}
}

