% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecasting from an object}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\description{
The functions allow producing forecasts based on the provided object.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("forecast")}
}

