% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate values based on inputs}
\usage{
generate(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods}
}
\description{
Generate values based on inputs
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("generate")}
}

