% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{generic_test}
\alias{generic_test}
\title{Generic test}
\usage{
generic_test(x)
}
\arguments{
\item{x}{A parameter}
}
\description{
Generic test
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("generic_test")}
}

