% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{glance}
\alias{glance}
\title{glance}
\usage{
glance(x, ...)
}
\arguments{
\item{x}{An object. See the individual method for specifics.}

\item{...}{Additional arguments to pass to other functions.}
}
\description{
glance
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("glance")}
}

