% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{hypothesize}
\alias{hypothesize}
\title{Construct hypotheses.}
\usage{
hypothesize(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods}
}
\description{
Construct hypotheses.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("hypothesize")}
}

