% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate missing values}
\usage{
interpolate(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Other arguments passed to methods}
}
\value{
A dataset (\code{\link[tibble:tibble]{tibble::tibble()}} or similar) of the same structure as
the input dataset with missing values from the response variable replaced
with interpolated values.
}
\description{
Interpolates missing values provided in the training dataset using the
fitted model.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("interpolate")}
}

