% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\title{The mean() generic}
\usage{
mean(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic for the \link[base]{mean} function
}
