% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars}
\alias{pars}
\title{Parameter Names}
\usage{
pars(x, ...)
}
\arguments{
\item{x}{A model or analysis object.}

\item{...}{Other arguments passed to methods}
}
\description{
Gets the parameter names for a model or analysis object.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("pars")}
}

