% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_results.R
\name{rank_results}
\alias{rank_results}
\title{Show relative rankings of different items}
\usage{
rank_results(x, ...)
}
\arguments{
\item{x}{An object that represents a collection of items.}

\item{...}{Other arguments passed to methods}
}
\description{
The functions returns a summary that provides relative ranks of different
items in a collection.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("rank_results")}
}

