% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{refit}
\alias{refit}
\title{refit}
\usage{
refit(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Additional arguments to pass to other functions.}
}
\description{
refit
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("refit")}
}

