% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_pkgs.R
\name{required_pkgs}
\alias{required_pkgs}
\title{Determine packages required by objects}
\usage{
required_pkgs(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods}
}
\value{
A character string of packages that are required.
}
\description{
Determine packages required by objects
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("required_pkgs")}
}

