% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_colnames.R
\name{rownames}
\alias{rownames}
\alias{rownames<-}
\alias{colnames}
\alias{colnames<-}
\title{The rownames() and colnames() generics}
\usage{
rownames(x, do.NULL = TRUE, prefix = "row")

rownames(x) <- value

colnames(x, do.NULL = TRUE, prefix = "col")

colnames(x) <- value
}
\arguments{
\item{x}{An R object}

\item{do.NULL}{logical. If FALSE and names are NULL, names are created.}

\item{prefix}{for created names}

\item{value}{a valid value for that component of \link[base]{dimnames}(x). For a
matrix or array this is either NULL or a character vector of non-zero length equal
to the appropriate dimension.}
}
\description{
Generics for \code{rownames}/\code{\link[base]{colnames}}
}
\note{
Dispatch on the 1st arg (the 'x' arg) only!
}
