% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{sd}
\alias{sd}
\title{Standard Deviation Generic}
\usage{
sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or an R object which is coercible to one by as.double(x)}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Generic for the \link[stats]{sd} function
}
\note{
Dispatches only on 'x'
}
