% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{select}
\alias{select}
\title{Select Generic}
\usage{
select(.data, ...)
}
\arguments{
\item{...}{Additional arguments, for use in specific methods.}

\item{x}{An object.}
}
\description{
Select Generic
}

