% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify}
\alias{specify}
\title{Specify variables or other quantities.}
\usage{
specify(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods}
}
\description{
Specify variables or other quantities.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("specify")}
}

