% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var}
\alias{var}
\alias{cov}
\alias{cor}
\title{Correlation, Variance, and Covariance Generics}
\usage{
var(x, y = NULL, na.rm = FALSE, use)

cov(x, y = NULL, use = "everything", method = c("pearson", "kendall",
  "spearman"))

cor(x, y = NULL, use = "everything", method = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{NULL (default) or a vector, matrix or data frame with compatible dimensions
to x. The default is equivalent to y = x (but more efficient).}

\item{na.rm}{logical. Should missing values be removed?}

\item{use}{an optional character string giving a method for computing covariances
in the presence of missing values. This must be (an abbreviation of) one of the
strings "everything", "all.obs", "complete.obs", "na.or.complete", or
"pairwise.complete.obs".}

\item{method}{a character string indicating which correlation coefficient (or covariance)
is to be computed. One of "pearson" (default), "kendall", or "spearman": can be
abbreviated.}
}
\description{
Generics for the \code{var}, \code{cov},
\code{\link[stats]{cor}} functions
}
\note{
Dispatches only on 'x' and 'y'
}
