% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{var_imp}
\alias{var_imp}
\title{var_imp}
\usage{
var_imp(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
var_imp
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("var_imp")}
}

