% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varying_args.R
\name{varying_args}
\alias{varying_args}
\title{Find any arguments that are not fully specified.}
\usage{
varying_args(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Other arguments passed to methods}
}
\description{
Find any arguments that are not fully specified.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("varying_args")}
}

