% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize a data set or object.}
\usage{
visualize(data, ...)
}
\arguments{
\item{data}{A data frame or other object.}

\item{...}{Other arguments passed to methods}
}
\description{
Visualize a data set or object.
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("visualize")}
}

