% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment}
\alias{augment}
\title{Augment data with information from an object}
\usage{
augment(x, data, ...)
}
\arguments{
\item{x}{Model object or other R object with information to append to
observations.}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}.}

\item{...}{Addition arguments to \code{augment} method.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about data points.
}
\description{
Augment data with information from an object
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("augment")}
}

