% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Configure an object}
\usage{
compile(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Other arguments passed to methods}
}
\description{
Finalizes or completes an object.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("compile")}
}

