% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components}
\alias{components}
\title{Extract components}
\usage{
components(object, ...)
}
\arguments{
\item{object}{A data separable object.}

\item{...}{Other arguments passed to methods}
}
\value{
A dataset (\code{\link[tibble:tibble]{tibble::tibble()}} or similar) containing components from
the object.
}
\description{
\code{components} can be used to extract elements from an object.
}
\details{
For example, decomposition methods and some modelling techniques can be used
to decompose a dataset into components of interest. This function is used
to extract these components in a tidy data format.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("components")}
}

