% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{estfun}
\alias{estfun}
\title{estfun}
\usage{
estfun(x, ...)
}
\arguments{
\item{x}{Model object or other R object with information to append to
observations.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
estfun
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("estfun")}
}

