% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.R
\name{estfun}
\alias{estfun}
\title{Extracting the estimating functions of a fitted model.}
\usage{
estfun(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Other arguments passed to methods}
}
\description{
Extracting the estimating functions of a fitted model.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("estfun")}
}

