% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(x, ...)
}
\arguments{
\item{x}{Model object or other R object with information to append to
observations.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
evaluate
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("evaluate")}
}

