% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate an object.}
\usage{
evaluate(x, ...)
}
\arguments{
\item{x}{An object. See the individual method for specifics.}

\item{...}{other arguments passed to methods}
}
\description{
Evaluate an object.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("evaluate")}
}

