% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain details of an object}
\usage{
explain(x, ...)
}
\arguments{
\item{x}{An object. See the individual method for specifics.}

\item{...}{other arguments passed to methods}
}
\description{
Explain details of an object
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("explain")}
}

