% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Additional arguments to pass to other functions.}
}
\description{
fit
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("fit")}
}

