% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Estimate model parameters.}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Other arguments passed to methods}
}
\description{
Estimates parameters for a given model from a set of data.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("fit")}
}

