% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xy.R
\name{fit_xy}
\alias{fit_xy}
\title{Estimate model parameters.}
\usage{
fit_xy(object, ...)
}
\arguments{
\item{object}{An object. See the individual method for specifics.}

\item{...}{Other arguments passed to methods}
}
\description{
Estimates parameters for a given model from a set of data in the form of
a set of predictors (\code{x}) and outcome(s) (\code{y}).
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("fit_xy")}
}

