% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fit}
\alias{fit}
\alias{varImp}
\alias{train}
\alias{learn}
\alias{object}
\alias{evaluate}
\title{Generic Functions Related to Modeling}
\usage{
fit(object, ...)

varImp(object, ...)

train(x, ...)

learn(x, ...)

object(x, ...)

evaluate(x, ...)
}
\arguments{
\item{object, x}{A objects.}

\item{...}{Additional arguments to pass to other funcitons.}
}
\description{
These are generic functions that can be used to minimize
package dependencies when multiple packages have the same
method.
}
\examples{
fit

train
}
