% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance}
\alias{glance}
\title{Glance at an object}
\usage{
glance(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\description{
Construct a single row summary "glance" of a model, fit, or other
object
}
\details{
glance methods always return either a one-row data frame (except on
\code{NULL}, which returns an empty data frame)
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("glance")}
}

