% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{learn}
\alias{learn}
\title{learn}
\usage{
learn(x, ...)
}
\arguments{
\item{x}{Model object or other R object with information to append to
observations.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
learn
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("learn")}
}

