% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{prune}
\alias{prune}
\title{prune}
\usage{
prune(tree, ...)
}
\arguments{
\item{tree}{A fitted model object.}

\item{...}{Addition arguments to \code{augment} method.}
}
\description{
prune
}
\section{Methods}{

\Sexpr[stage=render,results=Rd]{generics:::methods_rd("prune")}
}

