% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Prune or reduce an object}
\usage{
prune(tree, ...)
}
\arguments{
\item{tree}{A fitted model object.}

\item{...}{Other arguments passed to methods}
}
\description{
Prune or reduce an object
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("prune")}
}

