% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_results.R
\name{rank_results}
\alias{rank_results}
\title{Compute relative rankings of a collection of objects}
\usage{
rank_results(x, ...)
}
\arguments{
\item{x}{A collection of objects}

\item{...}{Other arguments passed to methods}
}
\description{
\code{rank_results()} computes relative ranks of a collection of objects and
returns a summary of the results.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("rank_results")}
}

