% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.R
\name{refit}
\alias{refit}
\title{Refitting models}
\usage{
refit(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Other arguments passed to methods}
}
\description{
Refitting models
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("refit")}
}

