% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Estimate model parameters.}
\usage{
train(x, ...)
}
\arguments{
\item{x}{An object. See the individual method for specifics.}

\item{...}{other arguments passed to methods}
}
\description{
Estimates parameters for a given model from a set of data.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("train")}
}

