% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.R
\name{var_imp}
\alias{var_imp}
\title{Calculation of variable importance}
\usage{
var_imp(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Other arguments passed to methods}
}
\description{
A generic method for calculating variable importance for model objects.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("var_imp")}
}

